refactor: upload dialog splitted into smaller widgets
