--- catalyst_voices/apps/voices/lib/widgets/modals/voices_upload_file_dialog.dart
@@ -1,354 +1,0 @@
-import 'package:catalyst_voices/dependency/dependencies.dart';
-import 'package:catalyst_voices/widgets/buttons/voices_filled_button.dart';
-import 'package:catalyst_voices/widgets/buttons/voices_outlined_button.dart';
-import 'package:catalyst_voices/widgets/indicators/voices_linear_progress_indicator.dart';
-import 'package:catalyst_voices/widgets/modals/voices_desktop_dialog.dart';
-import 'package:catalyst_voices/widgets/modals/voices_dialog.dart';
-import 'package:catalyst_voices_assets/catalyst_voices_assets.dart';
-import 'package:catalyst_voices_brands/catalyst_voices_brands.dart';
-import 'package:catalyst_voices_localization/catalyst_voices_localization.dart';
-import 'package:catalyst_voices_services/catalyst_voices_services.dart';
-import 'package:catalyst_voices_shared/catalyst_voices_shared.dart';
-import 'package:cross_file/cross_file.dart';
-import 'package:dotted_border/dotted_border.dart';
-import 'package:flutter/material.dart';
-import 'package:flutter_dropzone/flutter_dropzone.dart';
-
-typedef OnVoicesFileUploaded = Future<void> Function(XFile value);
-
-class VoicesUploadFileDialog extends StatefulWidget {
-  final String title;
-  final String itemNameToUpload;
-  final String? info;
-  final List<String>? allowedExtensions;
-  final OnVoicesFileUploaded? onUpload;
-
-  const VoicesUploadFileDialog({
-    super.key,
-    required this.title,
-    required this.itemNameToUpload,
-    this.info,
-    this.allowedExtensions,
-    this.onUpload,
-  });
-
-  @override
-  State<VoicesUploadFileDialog> createState() => _VoicesUploadFileDialogState();
-
-  static Future<XFile?> show(
-    BuildContext context, {
-    required RouteSettings routeSettings,
-    required String title,
-    String? itemNameToUpload,
-    String? info,
-    List<String>? allowedExtensions,
-    OnVoicesFileUploaded? onUpload,
-  }) {
-    return VoicesDialog.show<XFile?>(
-      context: context,
-      routeSettings: routeSettings,
-      builder: (context) {
-        return VoicesUploadFileDialog(
-          title: title,
-          itemNameToUpload: itemNameToUpload ?? context.l10n.file,
-          info: info,
-          allowedExtensions: allowedExtensions,
-          onUpload: onUpload,
-        );
-      },
-    );
-  }
-}
-
-class _Buttons extends StatefulWidget {
-  final XFile? selectedFile;
-  final OnVoicesFileUploaded? onUpload;
-
-  const _Buttons({
-    this.selectedFile,
-    this.onUpload,
-  });
-
-  @override
-  State<_Buttons> createState() => _ButtonsState();
-}
-
-class _ButtonsState extends State<_Buttons> {
-  bool _isUploading = false;
-
-  @override
-  Widget build(BuildContext context) {
-    return Row(
-      children: [
-        Expanded(
-          child: VoicesOutlinedButton(
-            onTap: () {
-              Navigator.of(context).pop();
-            },
-            child: Text(context.l10n.cancelButtonText),
-          ),
-        ),
-        if (!_isUploading) const SizedBox(width: 8),
-        if (!_isUploading)
-          Expanded(
-            child: VoicesFilledButton(
-              onTap: (widget.selectedFile != null)
-                  ? () async {
-                      setState(() {
-                        _isUploading = true;
-                      });
-                      await widget.onUpload?.call(widget.selectedFile!);
-
-                      if (context.mounted) {
-                        Navigator.pop(context, widget.selectedFile);
-                      }
-                    }
-                  : null,
-              child: Text(context.l10n.upload),
-            ),
-          ),
-      ],
-    );
-  }
-}
-
-class _InfoContainer extends StatelessWidget {
-  final String selectedFilename;
-  final bool isUploading;
-
-  const _InfoContainer({
-    required this.selectedFilename,
-    required this.isUploading,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      height: 72,
-      alignment: Alignment.centerLeft,
-      decoration: BoxDecoration(
-        borderRadius: BorderRadius.circular(12),
-        border: Border.all(
-          color: Theme.of(context).colors.iconsPrimary,
-        ),
-      ),
-      child: Row(
-        children: [
-          Padding(
-            padding: const EdgeInsets.symmetric(horizontal: 16),
-            child: VoicesAssets.icons.documentAdd.buildIcon(
-              color: Theme.of(context).colors.iconsPrimary,
-              size: 30,
-            ),
-          ),
-          if (isUploading)
-            Flexible(
-              child: Wrap(
-                direction: Axis.vertical,
-                children: [
-                  Text(
-                    context.l10n.uploadProgressInfo,
-                    style: Theme.of(context).textTheme.bodyMedium,
-                  ),
-                  const SizedBox(height: 8),
-                  const SizedBox(
-                    width: 480,
-                    child: VoicesLinearProgressIndicator(),
-                  ),
-                ],
-              ),
-            )
-          else
-            Text(
-              selectedFilename,
-              style: Theme.of(context).textTheme.bodyMedium,
-            ),
-        ],
-      ),
-    );
-  }
-}
-
-class _Title extends StatelessWidget {
-  final String title;
-
-  const _Title(this.title);
-
-  @override
-  Widget build(BuildContext context) {
-    return Text(
-      title.toUpperCase(),
-      style: Theme.of(context).textTheme.titleLarge,
-    );
-  }
-}
-
-class _UploadContainer extends StatefulWidget {
-  final String itemNameToUpload;
-  final String? info;
-  final List<String>? allowedExtensions;
-  final ValueChanged<XFile>? onFileSelected;
-
-  const _UploadContainer({
-    required this.itemNameToUpload,
-    this.info,
-    this.allowedExtensions,
-    this.onFileSelected,
-  });
-
-  @override
-  State<_UploadContainer> createState() => _UploadContainerState();
-}
-
-class _UploadContainerState extends State<_UploadContainer> {
-  late DropzoneViewController _dropzoneController;
-
-  @override
-  Widget build(BuildContext context) {
-    return Expanded(
-      child: Padding(
-        padding: const EdgeInsets.only(bottom: 12),
-        child: DottedBorder(
-          options: RoundedRectDottedBorderOptions(
-            radius: const Radius.circular(12),
-            dashPattern: const [8, 6],
-            color: Theme.of(context).colors.iconsPrimary,
-          ),
-          child: Stack(
-            children: [
-              // We allow drag&drop only on web
-              if (CatalystPlatform.isWeb)
-                DropzoneView(
-                  operation: DragOperation.copy,
-                  cursor: CursorType.grab,
-                  onCreated: (DropzoneViewController ctrl) => setState(() {
-                    _dropzoneController = ctrl;
-                  }),
-                  onDropFile: (ev) async {
-                    final bytes = await _dropzoneController.getFileData(ev);
-                    final name = await _dropzoneController.getFilename(ev);
-                    widget.onFileSelected?.call(
-                      XFile.fromData(
-                        bytes,
-                        name: name,
-                      ),
-                    );
-                  },
-                ),
-              InkWell(
-                onTap: () async {
-                  final service = Dependencies.instance.get<UploaderService>();
-                  final pickedFile = await service.uploadFile(
-                    allowedExtensions: widget.allowedExtensions,
-                  );
-                  if (pickedFile != null) {
-                    widget.onFileSelected?.call(pickedFile);
-                  }
-                },
-                child: Container(
-                  width: double.infinity,
-                  height: double.infinity,
-                  alignment: Alignment.center,
-                  child: Wrap(
-                    direction: Axis.vertical,
-                    crossAxisAlignment: WrapCrossAlignment.center,
-                    children:
-                        [
-                              Container(
-                                width: 80,
-                                height: 80,
-                                padding: const EdgeInsets.all(20),
-                                decoration: BoxDecoration(
-                                  shape: BoxShape.circle,
-                                  border: Border.all(
-                                    color: Theme.of(context).colors.iconsPrimary,
-                                    width: 3,
-                                  ),
-                                ),
-                                child: VoicesAssets.icons.upload.buildIcon(
-                                  color: Theme.of(context).colors.iconsPrimary,
-                                ),
-                              ),
-                              RichText(
-                                text: TextSpan(
-                                  text: context.l10n.uploadDropInfo(widget.itemNameToUpload),
-                                  style: Theme.of(context).textTheme.titleSmall,
-                                  children: <TextSpan>[
-                                    TextSpan(
-                                      text: context.l10n.browse,
-                                      style: TextStyle(
-                                        color: Theme.of(context).colors.iconsPrimary,
-                                      ),
-                                    ),
-                                  ],
-                                ),
-                              ),
-                              if (widget.info != null)
-                                Text(
-                                  widget.info!,
-                                  style: Theme.of(context).textTheme.bodyMedium,
-                                ),
-                            ]
-                            .separatedBy(
-                              const SizedBox(
-                                height: 12,
-                              ),
-                            )
-                            .toList(),
-                  ),
-                ),
-              ),
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-class _VoicesUploadFileDialogState extends State<VoicesUploadFileDialog> {
-  XFile? _selectedFile;
-  bool _isUploading = false;
-
-  @override
-  Widget build(BuildContext context) {
-    return VoicesSinglePaneDialog(
-      constraints: const BoxConstraints(maxWidth: 600, maxHeight: 450),
-      child: Padding(
-        padding: const EdgeInsets.all(16),
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            _Title(widget.title),
-            const SizedBox(height: 24),
-            _UploadContainer(
-              itemNameToUpload: widget.itemNameToUpload,
-              info: widget.info,
-              allowedExtensions: widget.allowedExtensions,
-              onFileSelected: (file) {
-                setState(() {
-                  _selectedFile = file;
-                });
-              },
-            ),
-            if (_selectedFile != null)
-              _InfoContainer(
-                selectedFilename: _selectedFile!.name,
-                isUploading: _isUploading,
-              ),
-            const SizedBox(height: 24),
-            _Buttons(
-              selectedFile: _selectedFile,
-              onUpload: (file) async {
-                setState(() {
-                  _isUploading = true;
-                });
-                await widget.onUpload?.call(file);
-              },
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
