+++ catalyst_voices/apps/voices/lib/widgets/widgets.dart
@@ -62,12 +62,12 @@ export 'menu/voices_node_menu.dart';
 export 'menu/voices_raw_popup_menu_button.dart';
 export 'modals/account/email_not_verified_dialog.dart';
 export 'modals/details/voices_details_dialog.dart';
+export 'modals/upload/voices_upload_file_dialog.dart';
 export 'modals/voices_alert_dialog.dart';
 export 'modals/voices_desktop_dialog.dart';
 export 'modals/voices_dialog.dart';
 export 'modals/voices_info_dialog.dart';
 export 'modals/voices_question_dialog.dart';
-export 'modals/voices_upload_file_dialog.dart';
 export 'rich_text/markdown_text.dart';
 export 'rich_text/placeholder_rich_text.dart';
 export 'scrollbar/voices_scrollbar.dart';
